% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACE_NIR.R
\name{np.export.date}
\alias{np.export.date}
\title{Merge}
\usage{
np.export.date(spl = lnir$S_115)
}
\arguments{
\item{spl}{csv file}
}
\value{
data.frame
}
\description{
recupere les metadata et les data NIRS puis fait la correspondence
}
\examples{
# blabla exemple
}
