% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACE_NIR.R
\name{np.import}
\alias{np.import}
\title{import from dx}
\usage{
np.import(name = "sample.dx")
}
\arguments{
\item{name}{csv file}
}
\value{
a list of lists
}
\description{
importe les donnees de l'instrument au format dx
}
\examples{
# np.import("sample.dx")
}
