% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACE_NIR.R
\name{np.read.block}
\alias{np.read.block}
\title{read block}
\usage{
np.read.block(jdx = lnir[[1]])
}
\arguments{
\item{jdx}{file jdx}
}
\value{
a data.frame
}
\description{
lit les block du fichier et les transforment au bon format
}
\examples{
# blabla exemple
}
