% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PACE_NIR.R
\name{np.split.Multiblock.DX}
\alias{np.split.Multiblock.DX}
\title{import dx files}
\usage{
np.split.Multiblock.DX(file = NULL, wd = NULL, export = FALSE)
}
\arguments{
\item{file}{file csv.}

\item{wd}{working directory}

\item{export}{logical}
}
\value{
data.frame with data and or csv file
}
\description{
import et fait pleins d'autres choses
}
\examples{
# blabla exemple
}
